requirements.txt
django-ninja
django-cors-headers
django-ninja-extra
django-ninja-jwt
ninja-schema
django-unfold